Extract the Saml module (SAML core, validators, caches) into Api/Saml
Ninth module of the #777 folder migration — the SAML surface. Moves the 13
Saml* types into Jellyfin.Plugin.SSO_Auth.Api.Saml:

- SamlAssertionValidator, SamlAuthorizeStateBuilder, SamlCertificate,
  SamlInsecureToggles, SamlLoginPolicy, SamlOutcomeStore, SamlRecipientValidator,
  SamlRedirectSigner, SamlReplayCache, SamlRequestCache, SamlResponseLoader,
  SamlSigningKey, SamlSpMetadataBuilder -> Api/Saml/ (namespace Api.Saml), import
  added at each call site.

Saml may depend on Authz (role mapping) and RateLimit (throttling); it also
imports the still-flat Api core (SsoUrlBuilder, VerifiedIdentity) which the DAG
rule permits during the migration. Registers the Saml -> {Authz, RateLimit}
edges.

Two qualified references that the move broke are fixed rather than papered over
with a wrong-direction import:
- Saml.cs (root SAML library) used `Api.SamlRedirectSigner` -> now the
  using-resolved unqualified name.
- SamlSecondaryCertificateTests used a fully-qualified `...Api.SamlResponseLoader`
  -> `...Api.Saml.SamlResponseLoader`.
- The Authz leaf RolePrivilegeMapper carried four `<see cref="SamlLoginPolicy"/>`
  doc refs pointing UP at the Saml module; they are fully-qualified so the doc
  link survives without importing Saml into a leaf (which the DAG rule forbids).

Pure mechanical move, no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (1579 both TFMs), the module
DAG conformance rule, and the ABI-floor build.

Refs #777
