+++ SSO-Auth.Tests/SamlRequestCacheTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Globalization;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.Saml;
 using Xunit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Tests;
