+++ SSO-Auth/Api/Saml/SamlRecipientValidator.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Jellyfin.Plugin.SSO_Auth.Api;
+namespace Jellyfin.Plugin.SSO_Auth.Api.Saml;
 
 /// <summary>
 /// Pure endpoint-binding check for a SAML response (#156): the bearer SubjectConfirmationData
