+++ SSO-Auth/Config/ProviderConfigStore.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Jellyfin.Plugin.SSO_Auth.Api;
 using Jellyfin.Plugin.SSO_Auth.Api.Audit;
+using Jellyfin.Plugin.SSO_Auth.Api.Saml;
 using MediaBrowser.Model.Plugins;
 using Microsoft.Extensions.Logging;
