+++ SSO-Auth/Config/ProviderConfigValidator.cs
@@ -4,6 +4,7 @@ using Jellyfin.Plugin.SSO_Auth.Api;
 using Jellyfin.Plugin.SSO_Auth.Api.Authz;
 using Jellyfin.Plugin.SSO_Auth.Api.Net;
 using Jellyfin.Plugin.SSO_Auth.Api.Provider;
+using Jellyfin.Plugin.SSO_Auth.Api.Saml;
 
 namespace Jellyfin.Plugin.SSO_Auth.Config;
